#31 Move global components to /global directory
